move /utils.py contains only GUID() to openwifi.py
This fixes also openwifi/models.py which uses this GUID class
