+++ openwifi/guid.py
